requirements.txt
langgraph>=1.0.0
langchain>=1.0.0
langchain-community>=0.3.0
ollama>=0.4.0
flask>=3.0.0
flask-cors>=4.0.0
pytest>=8.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
